mise à jour en fonction des tests

+++ c-skeleton/headers/binary_file_reader.h
@@ -14,7 +14,7 @@ uint32_t get_dimension_from_binary_file(FILE* file);
 
 // Déclaration de la fonction pour obtenir le nombre de vecteurs à partir du fichier binaire
 
-uint32_t get_nbr_vectors_from_binary_file(FILE *file);
+uint64_t get_nbr_vectors_from_binary_file(FILE *file);
 
 //// Déclaration de la fonction pour obtenir les informations sur les coordonnées des points
